Split expression simple forms
